requirements-server.txt
flask
flask-cors
Flask-Caching
Flask-WTF
pysha3
pyjwt
eth_account
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
skyfield
cryptography
python-dateutil
coinaddr
bs4
pyyaml

# Drop-in replacement for configparser which also reads config files
ConfigArgParse